chore: service message test and structure
